requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
numba>=0.65.0
cupy-cuda12x>=13.0.0  # optional, GPU acceleration (requires NVIDIA GPU + CUDA Toolkit)
